Filebeat refactoring and cleanup
* Cleanup variable naming
* Update state also if lines are skipped / excluded
* Introduce option to update state without sending an event
