+++ filebeat/beater/publish.go
@@ -101,7 +101,10 @@ func (p *syncLogPublisher) Start() {
 
 			pubEvents := make([]common.MapStr, 0, len(events))
 			for _, event := range events {
-				pubEvents = append(pubEvents, event.ToMapStr())
+				// Only send event with bytes read. 0 Bytes means state update only
+				if event.Bytes > 0 {
+					pubEvents = append(pubEvents, event.ToMapStr())
+				}
 			}
 
 			p.client.PublishEvents(pubEvents, publisher.Sync, publisher.Guaranteed)
